Issue 669 method A (#670)
This PR solved many bugs in method A. 

Most of the bugs are related to the misunderstanding of ids from trajectories and ids from the matrix of the data. 

Therefore, files with discontinuous and/or not sorted ids lead to wrong results or even to segFaults.

Besides, in this PR jpsreport's support for xml trajectories has been deprecated.

Some trivial tests based on jpscore simulations (using sources)
are added. 

When agents start at some non-zero time, the result should be
(time to reach the line) + starting time.

For example:

- Pedestrian needs 7 seconds to reach the goal starts at time=0s.
 --> arrival time is 7s

- Pedestrian needs 9 seconds to reach the goal starts at time=1s.
 --> arrival time is 8s

fixes #669

Co-authored-by: David Haensel <david@davidscorner.de>