Fix mean velocity calculation
If not all peds were in a frame it could happen, that the access to the velocity vector was out of range of the vector (now checked with at instead of []). To avoid that the vector containing the velocities is now created at the beginning with a length of _maxID, which will be too large.